Fix critical serial member initialisation.
The assignment of the `Serial` instance in the constructor is ineffective because a reference can only be initialised, not assigned. Thus, the constructor fails to initialise the `serial_` member. See this for a great explanation: https://stackoverflow.com/a/8636184/3697870

I have changed it to a traditional initialiser list like in the official library, and dropped the const since it's incompatible with references in this form.